requirements.txt
flask
pdfplumber
requests
python-dotenv
gunicorn
